[win32] implement subprocess helpers
This change adds a Windows implementation for the SubProcess class.

We're also implementing "run_command" for Windows, re-using
the SubProcess class. It might be seen as just a convenience
wrapper on top of that. At the moment, the Linux bits are left
unchanged, although the same implementation might be used for all
platforms.